+++ priv/repo/migrations/20201130145206_alter_rule_results_counts.exs
@@ -1,0 +1,10 @@
+defmodule TdDq.Repo.Migrations.AlterRuleResultsCounts do
+  use Ecto.Migration
+
+  def change do
+    alter table("rule_results") do
+      modify(:errors, :bigint, from: :integer)
+      modify(:records, :bigint, from: :integer)
+    end
+  end
+end